--- rubrics-ui/src/app/edit-program/page.tsx
@@ -1,14 +1,0 @@
-import ProgramEdit from '@/components/Program/ProgramEdit/EditProgram';
-import { getSpaceServerSide } from '@/utils/space/getSpaceServerSide';
-
-async function EditProgram() {
-  const space = (await getSpaceServerSide())!;
-
-  return (
-    <div>
-      <ProgramEdit space={space} />
-    </div>
-  );
-}
-
-export default EditProgram;
