+++ rubrics-ui/src/components/Program/ProgramEdit/EditProgramFullScreenModal.tsx
@@ -18,17 +18,15 @@ const EditProgram: React.FC<EditProgramProps> = ({ programId, space }) => {
   const handleBack = () => {
     router.push(`/programs/`);
   };
-  const handleClose = () => {};
+
   return (
-    <FullScreenModal open={true} onClose={handleClose} title="Edit Program" showCloseButton={false}>
-      <div className="p-10 max-w-4xl mx-auto  rounded-lg shadow-lg">
-        <Button onClick={handleBack} className="flex items-center">
-          <ChevronLeftIcon className="h-5 w-5 mr-2" />
-          Programs
-        </Button>
-        <ProgramEditScreen programId={programId} space={space} />
-      </div>
-    </FullScreenModal>
+    <div>
+      <Button onClick={handleBack}>
+        <ChevronLeftIcon className="h-5 w-5 mr-2" />
+        Programs
+      </Button>
+      <ProgramEditScreen programId={programId} space={space} />
+    </div>
   );
 };
