+++ rubrics-ui/src/components/Program/ProgramEdit/ProgramEditScreen.tsx
@@ -1,20 +1,20 @@
 'use client';
 
-import React, { useEffect, useState } from 'react';
-import ProgramInput from '@/components/Program/ProgramEdit/ProgramInput';
-import FullPageModal from '@dodao/web-core/components/core/modals/FullPageModal';
+import ProgramForm from '@/components/Program/ProgramEdit/ProgramForm';
+import SelectRubricModal, { RubricSummary } from '@/components/Program/ProgramEdit/SelectRubricModal';
+import { SpaceWithIntegrationsFragment } from '@/types/rubricsTypes/types';
 import Button from '@dodao/web-core/components/core/buttons/Button';
-import { EditProgramRubricProps, SpaceWithIntegrationsFragment } from '@/types/rubricsTypes/types';
-import MarkdownEditor from '@/components/MarkdownEditor/MarkdownEditor';
 import { useNotificationContext } from '@dodao/web-core/ui/contexts/NotificationContext';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
+import React, { useEffect, useState } from 'react';
+
 interface EditProgramProps {
   programId?: string;
   space: SpaceWithIntegrationsFragment;
 }
 
 function EditProgram({ programId, space }: EditProgramProps) {
-  const [rubrics, setRubrics] = useState<EditProgramRubricProps[]>([]);
+  const [rubrics, setRubrics] = useState<RubricSummary[]>([]);
   const [showSelectRubricsModal, setShowSelectRubricsModal] = useState<boolean>(false);
   const [selectedRubrics, setSelectedRubrics] = useState<string[]>([]);
   const [newProgram, setNewProgram] = useState<{ name: string; details: string; summary: string }>({
@@ -91,8 +91,8 @@ function EditProgram({ programId, space }: EditProgramProps) {
   };
 
   return (
-    <div className="mt-10 p-2 flex flex-col items-center justify-center gap-6">
-      <ProgramInput newProgram={newProgram} setNewProgram={setNewProgram} error={error} setError={setError} />
+    <div>
+      <ProgramForm newProgram={newProgram} setNewProgram={setNewProgram} error={error} setError={setError} />
 
       <Button variant="contained" primary onClick={() => setShowSelectRubricsModal(true)} className="mt-2">
         Select Rubrics
@@ -100,7 +100,7 @@ function EditProgram({ programId, space }: EditProgramProps) {
 
       <Button
         variant="contained"
-        className="mt-2"
+        className="mt-2 ml-2"
         primary
         disabled={!newProgram.name || !newProgram.details || !newProgram.summary}
         onClick={handleSaveProgram}
@@ -108,23 +108,13 @@ function EditProgram({ programId, space }: EditProgramProps) {
         {programId ? 'Update Program' : 'Create Program'}
       </Button>
 
-      <FullPageModal open={showSelectRubricsModal} onClose={() => setShowSelectRubricsModal(false)} title="Select Rubrics">
-        <div className="p-4 flex flex-wrap gap-4">
-          {rubrics.map((rubric) => (
-            <div
-              key={rubric.id}
-              className={`border border-gray-300 rounded-lg p-4 cursor-pointer ${selectedRubrics.includes(rubric.id) ? 'border-blue-500' : ''}`}
-              onClick={() => handleRubricSelect(rubric.id)}
-            >
-              <h2 className="text-lg font-bold truncate">{rubric.name}</h2>
-              <p className="text-sm text-gray-600 truncate">{rubric.summary}</p>
-            </div>
-          ))}
-        </div>
-        <Button variant="contained" primary onClick={() => setShowSelectRubricsModal(false)}>
-          Done
-        </Button>
-      </FullPageModal>
+      <SelectRubricModal
+        open={showSelectRubricsModal}
+        onClose={() => setShowSelectRubricsModal(false)}
+        selectedRubricIds={selectedRubrics}
+        rubrics={rubrics}
+        handleRubricSelect={handleRubricSelect}
+      />
     </div>
   );
 }
